upgrade to therubyracer 0.10.x call semantics
In JavaScript, every function call has a `this`
context. In prior versions of therubyracer and
therubyrhino, this context was simply dropped on 
the floor and only the arguments were passed to
ruby procs.

This has been corrected, and now every embedded
lambda or proc receives the `this` object as its
first parameter.

This commit upgrades to that interface for all
rails helpers.